Improved formatting
Same as before. A few grammar and spelling errors have been fixed.

I re-worded a few sentences:
29. I'm not sure about this one. I think the urgency of the 'no's should
either be high or low, and it was kinda slow before, but looked like it
was trying to be faster. I slowed it down, but if you want to speed it
up, you should change it to something like "Nonono Kenta, that's
alright"
38. "but I'm different" feels a bit weird and isn't really necessary, so
I removed that to make the sentence flow more
57. "uh" is something someone who is more occupied trying to remember
something would say, rather than "um"
64. Similarly, there should be a longer because between the "hmm" and
the rest of the sentence, considering he's in deep thought.